framework/requirements.txt
openai-agents
PyYAML